% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{CA}
\alias{CA}
\title{CA fonction form FActominer package}
\usage{
CA(
  X,
  ncp = 5,
  row.sup = NULL,
  col.sup = NULL,
  quanti.sup = NULL,
  quali.sup = NULL,
  graph = TRUE,
  axes = c(1, 2),
  row.w = NULL,
  excl = NULL
)
}
\arguments{
\item{excl}{}
}
\description{
CA fonction form FActominer package
}
\examples{
result <- CA(Rhone)
}
