% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframe.R
\name{check_dataframe}
\alias{check_dataframe}
\title{Check if df appropriate for CA}
\usage{
check_dataframe(df)
}
\arguments{
\item{df}{that we hope to be appropriate for CA}
}
\value{
Boolean
}
\description{
Check if df appropriate for CA
}
\examples{
check_dataframe(Rhone)
}
