% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pptx.CA.R
\name{check_pptx.CA}
\alias{check_pptx.CA}
\title{Check officer and rvg before loading help functions
Without officer and rvg won't work!!}
\usage{
check_pptx.CA()
}
\description{
Check officer and rvg before loading help functions
Without officer and rvg won't work!!
}
