% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pptx.R
\name{check_pptx}
\alias{check_pptx}
\title{Check officer and rvg before loading help functions
Without officer and rvg won't work!!}
\usage{
check_pptx()
}
\description{
Check officer and rvg before loading help functions
Without officer and rvg won't work!!
}
