% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.CA.R
\name{create_CA}
\alias{create_CA}
\title{Create CA object}
\usage{
create_CA(
  file_csv,
  folder = NULL,
  title = NULL,
  row.sup = NULL,
  col.sup = NULL,
  graph = FALSE,
  ncp = 2
)
}
\arguments{
\item{file_csv}{csv file}

\item{folder}{default working directory}

\item{title}{default name of the file}

\item{row.sup}{if row.sup= 1 remove first line}

\item{col.sup}{if col.sup=5 means don't use 5th column}

\item{graph}{plot the CA graph from factominer}

\item{ncp}{default is 2}
}
\value{
Object class CA
}
\description{
Create a CA object leveraging the function CA from the package FactoMiner
By using row.sup and col.sup, we can exclude some row or column from the analysis
}
\examples{
file_csv <- system.file(package = "HelpMe", "extdata/child_red.csv")
##data_map <- read.csv(file_csv) if you want to check your file in R
result.ca <- HelpMe::create_CA(file_csv, folder = NULL)
print("Coordonnates")
result.ca$PM_coord
##HelpMe::export_CA(result.ca) #if you want to export the result
}
\author{
Leverage Factominer package
}
