% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.CA.R
\name{create_CA_from_df}
\alias{create_CA_from_df}
\title{Create CA object from df or file}
\usage{
create_CA_from_df(df, title = NULL, row.sup = NULL, graph = FALSE, ncp = 2)
}
\arguments{
\item{ncp}{}
}
\description{
Create CA object from df or file
}
