% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.CA.R
\name{create_CA_from_file}
\alias{create_CA_from_file}
\title{Create CA from csv file}
\usage{
create_CA_from_file(
  file_csv,
  folder = ".",
  title = NULL,
  row.sup = NULL,
  col.sup = NULL,
  graph = FALSE,
  ncp = 2
)
}
\arguments{
\item{col.sup}{}
}
\description{
Create CA from csv file
}
