% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{dml}
\alias{dml}
\title{from rvg packages}
\usage{
dml(
  code,
  ggobj = NULL,
  bg = "white",
  fonts = list(),
  pointsize = 12,
  editable = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
from rvg packages
}
