% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.CA.R
\name{export_CA}
\alias{export_CA}
\title{Export the resultat of create_CA
save the coordonnates and the data used to run the CA as csv files}
\usage{
export_CA(res.ca, folder = NULL, rdssaved = FALSE)
}
\arguments{
\item{res.ca}{object de class CA}

\item{folder}{path}
}
\description{
Export the resultat of create_CA
save the coordonnates and the data used to run the CA as csv files
}
\examples{
## result.ca <- HelpMe::create_CA(file_csv, folder = NULL)
## HelpMe::export.CA(result.ca, "output") if output folder exists!

}
