% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelpMe-package.R
\docType{package}
\name{HelpMe-package}
\alias{HelpMe}
\alias{HelpMe-package}
\title{HelpMe: The package Helps ME}
\description{
Several functions to help me.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/scalgary92/HelpMe}
  \item Report bugs at \url{https://github.com/scalgary92/HelpMe/issues}
}

}
\author{
\strong{Maintainer}: Sandrine Lebon \email{sandrine.lebon@ipsos.com}

}
\keyword{internal}
