% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.cleaning.names.R
\name{light.cleaning.names}
\alias{light.cleaning.names}
\title{Light cleaning}
\usage{
light.cleaning.names(df)
}
\arguments{
\item{df}{}
}
\description{
Light cleaning
}
\examples{
names(Rhone)
names(light.cleaning.names(Rhone))
}
