% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot_quadmap.R
\name{myplot_quadmap}
\alias{myplot_quadmap}
\title{Title}
\usage{
myplot_quadmap(
  inputData,
  isHE = FALSE,
  masterbrand = NULL,
  force = 10,
  max.overlaps = 10
)
}
\arguments{
\item{inputData}{data for quadmap 3 columns: label, performance and impact}

\item{isHE}{map for HE}

\item{masterbrand}{if it is HE}

\item{force}{argument for ggrepel}

\item{max.overlaps}{argument for ggrepel}
}
\description{
Title
}
