% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CA.R
\name{plot.CA}
\alias{plot.CA}
\title{Title}
\usage{
\method{plot}{CA}(
  x,
  axes = c(1, 2),
  xlim = NULL,
  ylim = NULL,
  invisible = c("none", "row", "col", "row.sup", "col.sup", "quali.sup"),
  choix = c("CA", "quanti.sup"),
  col.row = "blue",
  col.col = "red",
  col.row.sup = "darkblue",
  col.col.sup = "darkred",
  col.quali.sup = "magenta",
  col.quanti.sup = "blue",
  label = c("all", "none", "row", "row.sup", "col", "col.sup", "quali.sup"),
  title = NULL,
  palette = NULL,
  autoLab = c("auto", "yes", "no"),
  new.plot = FALSE,
  selectRow = NULL,
  selectCol = NULL,
  unselect = 0.7,
  shadowtext = FALSE,
  habillage = "none",
  legend = list(bty = "y", x = "topleft"),
  graph.type = c("ggplot", "classic"),
  ggoptions = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
