% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.file.R
\name{read.file}
\alias{read.file}
\title{Title}
\usage{
read.file(file_csv, folder = ".")
}
\arguments{
\item{file_csv}{nom fichier}

\item{folder}{folder could be omitted}
}
\value{
df
}
\description{
Title
}
