% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rhone}
\alias{Rhone}
\title{Test data}
\format{
\subsection{\code{Rhone}}{

A data frame with 10 rows and 8 columns:
}
}
\source{
moi
}
\usage{
Rhone
}
\description{
Report ...
}
\keyword{datasets}
