% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_quadmap.R
\name{run_quadmap}
\alias{run_quadmap}
\title{Title}
\usage{
run_quadmap(
  file_csv,
  folder = ".",
  isHE = FALSE,
  masterbrand = NULL,
  force = 10,
  max.overlaps = 10
)
}
\arguments{
\item{file_csv}{file name or could be path}

\item{folder}{path for file name}

\item{isHE}{is it an HE}

\item{masterbrand}{masterbrand pour HE}

\item{force}{argument pour ggrepel}

\item{max.overlaps}{argument pour ggrepel}
}
\description{
Title
}
