% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_quadmap.R
\name{run_quadmap_BG}
\alias{run_quadmap_BG}
\title{Title}
\usage{
run_quadmap_BG(
  file_csv,
  folder = NULL,
  isHE = FALSE,
  TBdetail = "T2B",
  masterbrand = NULL,
  force = 10,
  max.overlaps = 10
)
}
\arguments{
\item{max.overlaps}{}
}
\description{
Title
}
