% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd.triplet.R
\name{svd.triplet}
\alias{svd.triplet}
\title{svd.triplet from Factominer package}
\usage{
svd.triplet(X, row.w = NULL, col.w = NULL, ncp = Inf)
}
\arguments{
\item{ncp}{}
}
\description{
svd.triplet from Factominer package
}
