% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.CA.R
\name{export_CA}
\alias{export_CA}
\title{Export the resultat of create_CA
save a red object, the coordonnates and the data used to run the CA}
\usage{
export_CA(res.ca, folder = NULL)
}
\arguments{
\item{res.ca}{object de class CA}

\item{folder}{path}
}
\description{
Export the resultat of create_CA
save a red object, the coordonnates and the data used to run the CA
}
\author{
Leverage Factominer package
}
