% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.cleaning.names.R
\name{light.cleaning.names}
\alias{light.cleaning.names}
\title{Title}
\usage{
light.cleaning.names(df)
}
\arguments{
\item{df}{}
}
\value{
df with cleaner names columns
}
\description{
Title
}
\examples{
names(Rhone)
names(light.cleaning.names(Rhone))

}
